+++ BINDER/BINDER/Share/CustomFirebaseAPI.swift
@@ -1,0 +1,44 @@
+//
+//  CustomFirebaseAPI.swift
+//  BINDER
+//
+//  Created by 김가은 on 2022/05/01.
+//
+
+import Foundation
+import Firebase
+import FirebaseFirestore
+
+public func ShowScheduleList(type : String, date : String, datestr: String, scheduleTitles : [String], scheduleMemos : [String], count : Int) {
+    let db = Firestore.firestore()
+    
+    var varScheduleTitles = scheduleTitles
+    var varScheduleMemos = scheduleMemos
+    var varCount: Int = count
+    
+    // 데이터베이스에서 일정 리스트 가져오기
+    let docRef = db.collection(type).document(Auth.auth().currentUser!.uid).collection("schedule").document(date).collection("scheduleList")
+    // Date field가 현재 날짜와 동일한 도큐먼트 모두 가져오기
+    docRef.whereField("date", isEqualTo: datestr).getDocuments() { (querySnapshot, err) in
+        if let err = err {
+            print("Error getting documents: \(err)")
+        } else {
+            for document in querySnapshot!.documents {
+                print("\(document.documentID) => \(document.data())")
+                // 사용할 것들 가져와서 지역 변수로 저장
+                let scheduleTitle = document.data()["title"] as? String ?? ""
+                let scheduleMemo = document.data()["memo"] as? String ?? ""
+                
+                if (!scheduleTitles.contains(scheduleTitle)) {
+                    // 여러 개의 일정이 있을 수 있으므로 가져와서 배열에 저장
+                    varScheduleTitles.append(scheduleTitle)
+                    varScheduleMemos.append(scheduleMemo)
+                }
+                
+                // 일정의 제목은 필수 항목이므로 일정 제목 개수만큼을 개수로 지정
+                varCount = scheduleTitles.count
+            }
+        }
+    }
+    
+}
