+++ BINDER/BINDER/ViewController/Schedule/ScheduleListViewController.swift
@@ -46,32 +46,9 @@ class ScheduleListViewController: UIViewController {
         let date = formatter.date(from: dateWithoutDays[0])!
         let datestr = formatter.string(from: date)
         
-        
         // 데이터베이스에서 일정 리스트 가져오기
-        let docRef = self.db.collection(self.type).document(Auth.auth().currentUser!.uid).collection("schedule").document(self.date).collection("scheduleList")
-        // Date field가 현재 날짜와 동일한 도큐먼트 모두 가져오기
-        docRef.whereField("date", isEqualTo: datestr).getDocuments() { (querySnapshot, err) in
-            if let err = err {
-                print("Error getting documents: \(err)")
-            } else {
-                for document in querySnapshot!.documents {
-                    print("\(document.documentID) => \(document.data())")
-                    // 사용할 것들 가져와서 지역 변수로 저장
-                    let scheduleTitle = document.data()["title"] as? String ?? ""
-                    let scheduleMemo = document.data()["memo"] as? String ?? ""
-                    
-                    if (!self.scheduleTitles.contains(scheduleTitle)) {
-                        // 여러 개의 일정이 있을 수 있으므로 가져와서 배열에 저장
-                        self.scheduleTitles.append(scheduleTitle)
-                        print(scheduleTitle)
-                        self.scheduleMemos.append(scheduleMemo)
-                    }
-                    
-                    // 일정의 제목은 필수 항목이므로 일정 제목 개수만큼을 개수로 지정
-                    self.count = self.scheduleTitles.count
-                }
-            }
-        }
+        ShowScheduleList(type: self.type, date: self.date, datestr: datestr, scheduleTitles: scheduleTitles, scheduleMemos: scheduleMemos, count: self.count)
+        
         scheduleListTableView.reloadData()
     }
